Major refactor of fem module and more Gmsh API improvements.
Included support for automating a series of GmshSessions using a GmshTask object making it possible to merge fem with different order, mesh size etc..
