+++ src/ada/fem/containers.py
@@ -1,10 +1,25 @@
+from __future__ import annotations
+
 import logging
 from bisect import bisect_left
 from dataclasses import dataclass
 from itertools import chain, groupby
 from operator import attrgetter
+from typing import Iterable, List
 
 import numpy as np
+
+from ada.concepts.points import Node
+from ada.concepts.structural import Beam
+from ada.config import Settings
+
+from .common import Amplitude, Csys, FemBase
+from .constraints import Bc, Constraint
+from .elements import Connector, ConnectorSection, Elem, FemSection, Mass
+from .interactions import Interaction, InteractionProperty
+from .sets import FemSet
+from .steps import Step
+from .surfaces import Surface
 
 
 @dataclass
@@ -15,6 +30,525 @@ class COG:
     sh_mass: float
     bm_mass: float
     no_mass: float
+
+
+class FEM(FemBase):
+    """
+    A FEM representation of its parent Part
+
+    :param name: Name of analysis model
+    :param parent: Part object
+    :param metadata: Attached metadata
+    :type parent: ada.Part
+    """
+
+    def __init__(self, name=None, parent=None, metadata=None):
+        from ada.concepts.containers import Nodes
+
+        metadata = metadata if metadata is not None else dict()
+        metadata["sensor_data"] = dict()
+        metadata["info"] = dict()
+        super().__init__(name, metadata, parent)
+        self._nodes = Nodes(parent=self)
+        self._elements = FemElements(fem_obj=self)
+        self._sets = FemSets(fem_obj=self)
+        self._sections = FemSections(fem_obj=self)
+        self._bcs = []
+        self._masses = dict()
+        self._constraints = []
+        self._surfaces = dict()
+        self._amplitudes = dict()
+        self._steps = list()
+        self._connectors = dict()
+        self._connector_sections = dict()
+        self._springs = dict()
+        self._intprops = dict()
+        self._interactions = dict()
+        self._sensors = dict()
+        self._predefined_fields = dict()
+        self._subroutine = None
+        self._initial_state = None
+        self._lcsys = dict()
+
+    def add_elem(self, elem: Elem):
+        elem.parent = self
+        self.elements.add(elem)
+
+    def add_section(self, section: FemSection):
+        section.parent = self
+        self.sections.add(section)
+
+    def add_bc(self, bc: Bc):
+        if bc.name in [b.name for b in self._bcs]:
+            raise Exception('BC with name "{bc_id}" already exists'.format(bc_id=bc.name))
+        bc.parent = self
+        if bc.fem_set.parent is None:
+            # TODO: look over this implementation. Is this okay?
+            logging.error("Bc FemSet has no parent. Adding to self")
+            self.sets.add(bc.fem_set)
+
+        self._bcs.append(bc)
+
+    def add_mass(self, mass: Mass):
+        mass.parent = self
+        self._masses[mass.name] = mass
+
+    def add_set(
+        self,
+        fem_set: FemSet,
+        ids=None,
+        p=None,
+        vol_box=None,
+        vol_cyl=None,
+        single_member=False,
+        tol=1e-4,
+    ) -> FemSet:
+        """
+        Simple method that creates a set string based on a set name, node or element ids and adds it to the assembly str
+
+        :param fem_set: A fem set object
+        :param ids: List of integers
+        :param p: Single point (x,y,z)
+        :param vol_box: Search by quadratic volume. Where p is (xmin, ymin, zmin) and vol_box is (xmax, ymax, zmax)
+        :param vol_cyl: Search by cylindrical volume. Used together with p to find
+                        nodes within cylinder inputted by [radius, height, thickness]
+        :param single_member: Set True if you wish to keep only a single member
+        :param tol: Point Tolerances. Default is 1e-4
+        """
+        if ids is not None:
+            fem_set.add_members(ids)
+
+        fem_set.parent = self
+
+        def append_members(nodelist):
+            if single_member is True:
+                fem_set.add_members([nodelist[0]])
+            else:
+                fem_set.add_members(nodelist)
+
+        if fem_set.type == "nset":
+            if p is not None or vol_box is not None or vol_cyl is not None:
+                nodes = self.nodes.get_by_volume(p, vol_box, vol_cyl, tol)
+                if len(nodes) == 0 and self.parent is not None:
+                    assembly = self.parent.get_assembly()
+                    list_of_ps = assembly.get_all_subparts() + [assembly]
+                    for part in list_of_ps:
+                        nodes = part.fem.nodes.get_by_volume(p, vol_box, vol_cyl, tol)
+                        if len(nodes) > 0:
+                            fem_set.parent = part.fem
+                            append_members(nodes)
+                            part.fem.add_set(fem_set)
+                            return fem_set
+
+                    raise Exception(f'No nodes found for fem set "{fem_set.name}"')
+                elif nodes is not None and len(nodes) > 0:
+                    append_members(nodes)
+                else:
+                    raise Exception(f'No nodes found for femset "{fem_set.name}"')
+
+        self._sets.add(fem_set)
+        return fem_set
+
+    def add_step(self, step: Step) -> Step:
+        """Add an analysis step to the assembly"""
+        if len(self._steps) > 0:
+            if self._steps[-1].type != "eigenfrequency" and step.type == "complex_eig":
+                raise Exception(
+                    "complex eigenfrequency analysis step needs to follow eigenfrequency step. Check your input"
+                )
+        step.parent = self
+        self._steps.append(step)
+
+        return step
+
+    def add_interaction_property(self, int_prop: InteractionProperty) -> InteractionProperty:
+        int_prop.parent = self
+        self._intprops[int_prop.name] = int_prop
+        return int_prop
+
+    def add_interaction(self, interaction: Interaction) -> Interaction:
+        interaction.parent = self
+        self._interactions[interaction.name] = interaction
+        return interaction
+
+    def add_constraint(self, constraint: Constraint) -> Constraint:
+        constraint.parent = self
+        self._constraints.append(constraint)
+        return constraint
+
+    def add_lcsys(self, lcsys: Csys) -> Csys:
+        if lcsys.name in self._lcsys.keys():
+            raise ValueError("Local Coordinate system cannot have duplicate name")
+        lcsys.parent = self
+        self._lcsys[lcsys.name] = lcsys
+        return lcsys
+
+    def add_connector_section(self, connector_section: ConnectorSection):
+        connector_section.parent = self
+        self._connector_sections[connector_section.name] = connector_section
+
+    def add_connector(self, connector: Connector):
+        connector.parent = self
+        self._connectors[connector.name] = connector
+        connector.csys.parent = self
+        self.elements.add(connector)
+        self.add_set(FemSet(name=connector.name, members=[connector.id], set_type="elset"))
+
+    def add_rp(self, name, node: Node):
+        """Adds a reference point in assembly with a specific name"""
+        node.parent = self
+        self.nodes.add(node)
+        fem_set = self.add_set(FemSet(name, [node], "nset"))
+        return node, fem_set
+
+    def add_surface(self, surface: Surface):
+        surface.parent = self
+        self._surfaces[surface.name] = surface
+
+    def add_amplitude(self, amplitude: Amplitude):
+        amplitude.parent = self
+        self._amplitudes[amplitude.name] = amplitude
+
+    def add_sensor(self, name, point, comment, tol=1e-2):
+        """
+
+        :param name: Name of coordinate set
+        :param point: Sensor Coordinate
+        :param comment: Comment
+        :param tol:
+        """
+        fem_set = FemSet(name, [], "nset", metadata=dict(comment=comment))
+        self.add_set(fem_set, p=point, tol=tol, single_member=True)
+        if name in self._sensors.keys():
+            raise Exception("{} exists in sensor sets and will be overwritten. Please change name.".format(name))
+
+        self._sensors[name] = fem_set
+        # self._cad[name].add_shape(ada.cad.utils.make_sphere(point, tol), colour='red', transparency=0.5)
+
+    def add_predefined_field(self, pre_field):
+        """
+
+        :type pre_field: PredefinedField
+        """
+        pre_field.parent = self
+
+        self._predefined_fields[pre_field.name] = pre_field
+
+    def add_spring(self, spring):
+        """
+
+        :param spring:
+        :type spring: Spring
+        """
+
+        # self.elements.add(spring)
+
+        if spring.fem_set.parent is None:
+            self.sets.add(spring.fem_set)
+        self._springs[spring.name] = spring
+
+    def convert_ecc_to_mpc(self):
+        """
+
+        Converts beam offsets to MPC constraints
+        """
+        from ada import Node
+        from ada.core.utils import vector_length
+
+        edited_nodes = dict()
+        tol = Settings.point_tol
+
+        def build_mpc(fs):
+            """
+
+            :param fs:
+            :type fs: FemSection
+            :return:
+            """
+            if fs.offset is None or fs.type != "beam":
+                return
+            elem = fs.elset.members[0]
+            for n_old, ecc in fs.offset:
+                i = elem.nodes.index(n_old)
+                if n_old.id in edited_nodes.keys():
+                    n_new = edited_nodes[n_old.id]
+                    mat = np.eye(3)
+                    new_p = np.dot(mat, ecc) + n_old.p
+                    n_new_ = Node(new_p, parent=elem.parent)
+                    if vector_length(n_new_.p - n_new.p) > tol:
+                        elem.parent.nodes.add(n_new_, allow_coincident=True)
+                        m_set = FemSet(f"el{elem.id}_mpc{i + 1}_m", [n_new_], "nset")
+                        s_set = FemSet(f"el{elem.id}_mpc{i + 1}_s", [n_old], "nset")
+                        c = Constraint(
+                            f"el{elem.id}_mpc{i + 1}_co",
+                            "mpc",
+                            m_set,
+                            s_set,
+                            mpc_type="Beam",
+                            parent=elem.parent,
+                        )
+                        elem.parent.add_constraint(c)
+                        elem.nodes[i] = n_new_
+                        edited_nodes[n_old.id] = n_new_
+
+                    else:
+                        elem.nodes[i] = n_new
+                        edited_nodes[n_old.id] = n_new
+                else:
+                    mat = np.eye(3)
+                    new_p = np.dot(mat, ecc) + n_old.p
+                    n_new = Node(new_p, parent=elem.parent)
+                    elem.parent.nodes.add(n_new, allow_coincident=True)
+                    m_set = FemSet(f"el{elem.id}_mpc{i + 1}_m", [n_new], "nset")
+                    s_set = FemSet(f"el{elem.id}_mpc{i + 1}_s", [n_old], "nset")
+                    c = Constraint(
+                        f"el{elem.id}_mpc{i + 1}_co",
+                        "mpc",
+                        m_set,
+                        s_set,
+                        mpc_type="Beam",
+                        parent=elem.parent,
+                    )
+                    elem.parent.add_constraint(c)
+
+                    elem.nodes[i] = n_new
+                    edited_nodes[n_old.id] = n_new
+
+        list(map(build_mpc, filter(lambda x: x.offset is not None, self.sections)))
+
+    def convert_hinges_2_couplings(self):
+        """
+        Convert beam hinges to coupling constraints
+        """
+        from ada import Node
+
+        def converthinges(fs):
+            """
+
+            :param fs:
+            :type fs: ada.fem.FemSection
+            """
+            if fs.hinges is None or fs.type != "beam":
+                return
+            elem = fs.elset.members[0]
+            assert isinstance(elem, Elem)
+
+            for n, d, csys in fs.hinges:
+                n2 = Node(n.p, None, parent=elem.parent)
+                elem.parent.nodes.add(n2, allow_coincident=True)
+                i = elem.nodes.index(n)
+                elem.nodes[i] = n2
+                if elem.fem_sec.offset is not None:
+                    if n in [x[0] for x in elem.fem_sec.offset]:
+                        elem.fem_sec.offset[i] = (n2, elem.fem_sec.offset[i][1])
+
+                s_set = FemSet(f"el{elem.id}_hinge{i + 1}_s", [n], "nset")
+                m_set = FemSet(f"el{elem.id}_hinge{i + 1}_m", [n2], "nset")
+                elem.parent.add_set(m_set)
+                elem.parent.add_set(s_set)
+                c = Constraint(
+                    f"el{elem.id}_hinge{i + 1}_co",
+                    "coupling",
+                    m_set,
+                    s_set,
+                    d,
+                    csys=csys,
+                )
+                elem.parent.add_constraint(c)
+
+        list(map(converthinges, filter(lambda x: x.hinges is not None, self.sections)))
+
+    def create_fem_elem_from_obj(self, obj, el_type=None) -> Elem:
+        """Converts structural object to FEM elements. Currently only BEAM is supported"""
+
+        if type(obj) is not Beam:
+            raise NotImplementedError(f'Object type "{type(obj)}" is not yet supported')
+
+        el_type = "B31" if el_type is None else el_type
+
+        res = self.nodes.add(obj.n1)
+        if res is not None:
+            obj.n1 = res
+        res = self.nodes.add(obj.n2)
+        if res is not None:
+            obj.n2 = res
+
+        elem = Elem(None, [obj.n1, obj.n2], el_type)
+        self.add_elem(elem)
+        femset = FemSet(f"{obj.name}_set", [elem.id], "elset")
+        self.add_set(femset)
+        self.add_section(
+            FemSection(
+                f"d{obj.name}_sec",
+                "beam",
+                femset,
+                obj.material,
+                obj.section,
+                obj.ori[1],
+            )
+        )
+        return elem
+
+    @property
+    def parent(self):
+        """
+
+        :rtype: ada.Part
+        """
+        return self._parent
+
+    @parent.setter
+    def parent(self, value):
+        from ada import Part
+
+        if issubclass(type(value), Part) is False and value is not None:
+            raise ValueError()
+        self._parent = value
+
+    @property
+    def nodes(self):
+        """
+
+        :rtype: ada.concepts.containers.Nodes
+        """
+        return self._nodes
+
+    @property
+    def elements(self) -> FemElements:
+        return self._elements
+
+    @property
+    def sections(self) -> FemSections:
+        return self._sections
+
+    @property
+    def bcs(self):
+        return self._bcs
+
+    @property
+    def constraints(self):
+        return self._constraints
+
+    @property
+    def instance_name(self):
+        return self._name if self._name is not None else f"{self.parent.name}-1"
+
+    @property
+    def sets(self):
+        """
+
+        :return:
+        :rtype: ada.fem.containers.FemSets
+        """
+        return self._sets
+
+    @property
+    def nsets(self):
+        return self.sets.nodes
+
+    @property
+    def elsets(self):
+        return self.sets.elements
+
+    @property
+    def masses(self):
+        return self._masses
+
+    @property
+    def interactions(self):
+        """
+
+        :rtype: dict
+        """
+        return self._interactions
+
+    @property
+    def intprops(self):
+        return self._intprops
+
+    @property
+    def steps(self):
+        """
+
+        :return:
+        :rtype: list
+        """
+        return self._steps
+
+    @property
+    def surfaces(self):
+        """
+
+        :rtype: dict
+        """
+        return self._surfaces
+
+    @property
+    def connectors(self):
+        return self._connectors
+
+    @property
+    def connector_sections(self):
+        return self._connector_sections
+
+    @property
+    def amplitudes(self):
+        return self._amplitudes
+
+    @property
+    def sensors(self):
+        return self._sensors
+
+    @property
+    def predefined_fields(self):
+        return self._predefined_fields
+
+    @property
+    def initial_state(self):
+        """
+
+        :return:
+        :rtype: PredefinedField
+        """
+        return self._initial_state
+
+    @initial_state.setter
+    def initial_state(self, value):
+        self._initial_state = value
+
+    @property
+    def springs(self):
+        return self._springs
+
+    @property
+    def lcsys(self):
+        return self._lcsys
+
+    def __add__(self, other: FEM):
+        # Nodes
+        nodid_max = self.nodes.max_nid if len(self.nodes) > 0 else 0
+        if nodid_max > other.nodes.min_nid:
+            other.nodes.renumber(int(nodid_max + 10))
+
+        self._nodes += other.nodes
+
+        # Elements
+        elid_max = self.elements.max_el_id if len(self.elements) > 0 else 0
+
+        if elid_max > other.elements.min_el_id:
+            other.elements.renumber(int(elid_max + 10))
+
+        logging.info("FEM operand type += is still ")
+
+        self._elements += other.elements
+        self._sections += other._sections
+        self._sets += other._sets
+        self._lcsys.update(other.lcsys)
+
+        return self
+
+    def __repr__(self):
+        return f"FEM({self.name}, Elements: {len(self.elements)}, Nodes: {len(self.nodes)})"
 
 
 class FemElements:
@@ -38,18 +572,12 @@ class FemElements:
 
         self._group_by_types()
 
-    def renumber(self):
+    def renumber(self, start_id=1):
         from ada.core.utils import Counter
 
-        elid = Counter(1)
+        elid = Counter(start_id)
 
-        def mapid2(el):
-            """
-
-            :param el:
-            :type el: ada.fem.Elem
-            :return:
-            """
+        def mapid2(el: Elem):
             el.id = next(elid)
 
         list(map(mapid2, self._elements))
@@ -266,6 +794,10 @@ class FemElements:
         return max(self._idmap.keys())
 
     @property
+    def min_el_id(self):
+        return min(self._idmap.keys())
+
+    @property
     def elements(self):
         return self._elements
 
@@ -312,39 +844,8 @@ class FemElements:
         return filter(lambda x: x.type in Mass._valid_types, self._elements)
 
     @property
-    def edges(self):
-        """
-
-
-        :return: List of edges for visualization for all elements
-        """
-
-        def grab_nodes(elem):
-            return [self._fem_obj.nodes.from_id(elem.nodes[e].id).p for ed_seq in elem.edges_seq for e in ed_seq]
-
-        return list(
-            chain.from_iterable(
-                map(
-                    grab_nodes,
-                    filter(lambda x: x.type not in ["MASS", "SPRING1"], self._elements),
-                )
-            )
-        )
-
-    @property
-    def edges_alt(self):
-        """
-
-        :return: List of edges for visualization for all elements (alternative implementation)
-        """
-
-        def grab_ids(elem):
-            return tuple([elem.nodes[e].id for ed_seq in elem.edges_seq for e in ed_seq])
-
-        return map(
-            grab_ids,
-            filter(lambda x: x.type not in ["MASS", "SPRING1"], self._elements),
-        )
+    def stru_elements(self) -> Iterable[Elem]:
+        return filter(lambda x: x.type not in ["MASS", "SPRING1"], self._elements)
 
     def from_id(self, el_id):
         """
@@ -475,7 +976,8 @@ class FemSections:
         self._shells = by_types["shells"]
         self._solids = by_types["solids"]
         self._dmap = {e.name: e for e in self._sections} if len(self._sections) > 0 else dict()
-        if len(self._sections) > 0:
+
+        if len(self._sections) > 0 and fem_obj is not None:
             self._link_data()
 
     def _map_materials(self, fem_sec, mat_repo):
@@ -542,11 +1044,15 @@ class FemSections:
         result = self._sections[index]
         return FemSections(result) if isinstance(index, slice) else result
 
-    def __add__(self, other):
-        return FemSections(chain(self._sections, other._sections))
+    def __add__(self, other: FemSections):
+        return FemSections(chain(self._sections, other.sections))
 
     def __repr__(self):
         return f"FemSections(Beams: {len(self.lines)}, Shells: {len(self.shells)}, Solids: {len(self.solids)})"
+
+    @property
+    def sections(self) -> List[FemSection]:
+        return self._sections
 
     @property
     def lines(self):
