+++ src/ada/fem/io_meshio/reader.py
@@ -3,24 +3,17 @@ from itertools import chain
 import meshio
 
 from ada.concepts.containers import Nodes
+from ada.concepts.levels import Assembly, Part
+from ada.concepts.points import Node
 from ada.core.utils import Counter
 from ada.fem import FEM, Elem
 from ada.fem.containers import FemElements
 
-
-def meshio_read_fem(assembly, fem_file, fem_name=None):
-    """
-    Import a FEM file using the meshio package.
+from .common import meshio_to_ada_type
 
 
-    :param assembly: Assembly object
-    :param fem_file: Path to fem file
-    :param fem_name: Name of FEM model
-    :type assembly: ada.Assembly
-    """
-    from ada import Node, Part
-
-    from . import meshio_to_ada_type
+def meshio_read_fem(assembly: Assembly, fem_file, fem_name=None):
+    """Import a FEM file using the meshio package"""
 
     mesh = meshio.read(fem_file)
     name = fem_name if fem_name is not None else "Part-1"
